refactor: extract TransactionTable into folder with hook
